Fix quiz type definitions
